Create an empty Rust crate
A new Rust crate has been initialized in the repository. The dependency
to gdnative has already been added, and the crate has been configured to
build a dynamic library for Godot. Additionally, a basic logging setup
has been added so that we can read log messages inside the Godot editor.