Create the script for the player
A new script has been created for the player. The scripts extends the
native `KinematicBody2D` class, and has been attached to the player
scene.

Following the tutorial, another script has been created for actors in
general. In GDScript, the player script inherits from this generic actor
script. In Rust, this is not possible due to the lack of inheritance.
When we see how the actor script is used in later parts of the tutorial,
we will be able to decide how to deal with it in Rust. A potential
solution might be to turn it into a `Trait` with default implementations
for the shared behavior.