Move the player to the right (#8)
The most basic form of movement has been implemented for the player. It
simply moves them with a constant speed to the right.

As hypothesized earlier, the implementation for the `Actor` parent class
has been changed to a `Trait` in Rust so that its behavior can be shared
between Rust types.