Move the player
The first step towards controlling the player has been implemented.
Players can use the `A` or `Left` keys to move the player character to
the left, and `D` and `Right` to move them to the right.

The implementation for the movement has been fully moved from the
`Actor` trait to the player class. Once we know how the enemies will
move, we can move shared behavior back into the trait.