Disable enemies outside the view (#17)
A visibility enabler has been added to enemies so that they do not move
when they are outside the player's view. This prevents them from
accidentally falling to their death prematurely, and ensures that the
player will encounter enemies at the places where we drop them when
building the level.