Despawn enemies when player jumps on them (#19)
When the player jumps on an enemy, the enemy dies and gets despawned.
This is implemented using a collision shape on top of the enemy, and a
signal that is sent whenever something collides with this shape. Because
enemies can only collide with the player and the world, we know that
only the player can have hit the top of the enemy.